refactor: detailView로 분리
